requirements.txt
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2
fastapi[standard]
google-auth
python-dotenv